structural_bioinformatics/requirements.txt
biopython
numpy==1.24.4
